UPSTREAM: cfg80211: support loading regulatory database as firmware file
As the current regulatory database is only about 4k big, and already
difficult to extend, we decided that overall it would be better to
get rid of the complications with CRDA and load the database into the
kernel directly, but in a new format that is extensible.

The new file format can be extended since it carries a length field
on all the structs that need to be extensible.

In order to be able to request firmware when the module initializes,
move cfg80211 from subsys_initcall() to the later fs_initcall(); the
firmware loader is at the same level but linked earlier, so it can
be called from there. Otherwise, when both the firmware loader and
cfg80211 are built-in, the request will crash the kernel. We also
need to be before device_initcall() so that cfg80211 is available
for devices when they initialize.

BUG=b:128831966
TEST=None

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
(cherry picked from commit 007f6c5e6eb45c81ee89368a5f226572ae638831
https: //github.com/torvalds/linux.git)
Signed-off-by: Khasim Syed Mohammed <khasim.mohammed@linaro.org>
Change-Id: Ide4231c6f7be97d35dfb7f902be8b0a562a9ad26
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1530595
Reviewed-by: Zhihong Yu <zhihongyu@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Zhihong Yu <zhihongyu@chromium.org>
Tested-by: Zhihong Yu <zhihongyu@chromium.org>