Embryo of PMS405 SPMI regulator support
The PMS405 sports 5 SMPS and 13 LDO regulators, this embryo adds them
all to the spmi regulator driver, which results in a splat in the log
about unsupported versions.

We need S3 support merged, which is a BUCK of type HFS430, so this patch
starts adding that support, but all operations needs to be reviewed and
adapted to support this type of regulator.

We will have to do some research in order to figure out the LDOs, but a
potential path forward is to omit these in the initial patch upstream -
as it's only S3 we want at this point in time.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>