BACKPORT: FROMLIST: dt-bindings: PCI: qcom: Add QCS404 to the binding
The Qualcomm QCS404 platform contains a PCIe controller, add this to the
Qualcomm PCI binding document. The controller is the same version as the
one used in IPQ4019, but the PHY part is described separately, hence the
difference in clocks and resets.

(am from https://lore.kernel.org/patchwork/patch/1035793/)

BUG=b:128831966
TEST=None

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Signed-off-by: Khasim Syed Mohammed <khasim.mohammed@linaro.org>
Change-Id: I151761562eea7e76b6fe62fe74a2ce2199911ba1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1530671
Reviewed-by: Zhihong Yu <zhihongyu@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Zhihong Yu <zhihongyu@chromium.org>
Tested-by: Zhihong Yu <zhihongyu@chromium.org>