CHROMIUM: WAR: Bluetooth: Remove BT chip power off for every BT OFF
This fixes the BT ON issue like in boot up it is trying to
do device up and down and again it do device up. few times FW
download fialing, so we are blocking the device down and
second device up in bootup. So by the time  device bootup
BT device up and running.

BUG=b:128831966
TEST=None

Change-Id: I720cb53f94234888437879b8f4c31f0366722283
Signed-off-by: Harish Bandi <c-hbandi@codeaurora.org>
Signed-off-by: Mayank Chopra <mak.chopra@codeaurora.org>
Signed-off-by: Sumit Agrawal <agsumit@codeaurora.org>
Signed-off-by: Vikas Patel <vikpatel@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1535777
Reviewed-by: Zhihong Yu <zhihongyu@chromium.org>
Commit-Queue: Zhihong Yu <zhihongyu@chromium.org>
Tested-by: Zhihong Yu <zhihongyu@chromium.org>