CHROMIUM: arm64: dts: qcom: qcs404-mistral: Use buck for vdd_apc
In order to reach 1.4GHz the external FAN53526 is providing VDD_APC, os
override the default S3 from PMS405 as supply for the CPUs. The 1.4GHz
opp is defined and voltage values are adjusted to match valid voltage
steps of the FAN53526.

Finally the SPMI PMS405 S3 regulator is disabled.

BUG=b:124524751
TEST=Stepped through each frequency and observed voltage change

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Signed-off-by: Khasim Syed Mohammed <khasim.mohammed@linaro.org>
Change-Id: I4eb9f77e092f252df090babb5859ab3143b3da9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1530713
Reviewed-by: Zhihong Yu <zhihongyu@chromium.org>
Commit-Queue: Zhihong Yu <zhihongyu@chromium.org>
Tested-by: Zhihong Yu <zhihongyu@chromium.org>