UPSTREAM: nvmem: core: fix read buffer in place
When the bit_offset in the cell is zero, the pointer to the msb will
not be properly initialized (ie, will still be pointing to the first
byte in the buffer).

This being the case, if there are bits to clear in the msb, those will
be left untouched while the mask will incorrectly clear bit positions
on the first byte.

This commit also makes sure that any byte unused in the cell is
cleared.

Signed-off-by: Jorge Ramirez-Ortiz <jorge.ramirez-ortiz@linaro.org>
Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 2fe518fecb3a4727393be286db9804cd82ee2d91
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git)
Signed-off-by: Niklas Cassel <niklas.cassel@linaro.org>

BUG=b:129558664
TEST=Validate DVFS and schedutil governor

Signed-off-by: Praveen Koya <pkoya@codeaurora.org>
Change-Id: I76ba888248fd4fc7d4251ef7336017d337f09b9d
Signed-off-by: Vamsi Singamsetty <vamssi@codeaurora.org>
Signed-off-by: Khasim Syed Mohammed <khasim.mohammed@linaro.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1664371
Reviewed-by: Zhihong Yu <zhihongyu@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Zhihong Yu <zhihongyu@chromium.org>
Tested-by: Zhihong Yu <zhihongyu@chromium.org>