BACKPORT: mac80211: probe unexercised mesh links
The requirement for mesh link metric refreshing, is that from one
mesh point we be able to send some data frames to other mesh points
which are not currently selected as a primary traffic path, but which
are only 1 hop away. The absence of the primary path to the chosen node
makes it necessary to apply some form of marking on a chosen packet
stream so that the packets can be properly steered to the selected node
for testing, and not by the regular mesh path lookup.

Tested-by: Pradeep Kumar Chitrapu <pradeepc@codeaurora.org>
Signed-off-by: Rajkumar Manoharan <rmanohar@codeaurora.org>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>

(cherry picked from commit 8828f81ad4a2f4e89ebe6e7793c06ed767c31d53)

BUG=b:134764458
TEST=manually test and verified by sniffer on mistral
Change-Id: Ibd16c06ecab90f2246bb515ed703a8576eb438a9
Signed-off-by: Roy Luo <royluo@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1650690
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Julan Hsu <julanhsu@google.com>