CHROMIUM: ath10k: Avoid overriding beacon Tx for non-p2p ROC
If want to use remain-on-channel interface for off-channel scanning
with wmi 10.2.4/10.4 in AP mode, passing a huge dwell time as big as
10 secs (which was added to fix some issues in p2p) has a potential
to miss beacon Tx depending on where we are with respect to the next
TBTT by switching to the requested channel right away. Beacon miss
during off-channel may not be preferred especially when the requested
off-channel (ROC) duration is less than the beacon interval. Make the
work around of passing very high dwell time specific to p2p because
the work around is added for p2p issue. For non-p2p, Pass the requested
off-channel duration in dwell time. Also defer scan stop worker
by some more time to account for the time delay to complete the scan.

BUG=b:132147953
TEST=None

Change-Id: Iab98c5dc7b2eb637899a337f8af40b6995458a99
Signed-off-by: Vasanthakumar Thiagarajan <vthiagar@codeaurora.org>
Signed-off-by: Zhi Chen <zhichen@codeaurora.org>
Signed-off-by: Vamsi Singamsetty <vamssi@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1597058
Reviewed-by: Zhihong Yu <zhihongyu@chromium.org>
Reviewed-by: Govind Singh <govinds@qualcomm.corp-partner.google.com>
Commit-Queue: Zhihong Yu <zhihongyu@chromium.org>
Tested-by: Zhihong Yu <zhihongyu@chromium.org>