CHROMIUM: ath10k: Skip queuing restart work when wmi cmd timeout and firmware is down
The condition in wait_event_timeout is not atomic, so when timeout
happens, this may be lead to preemption in between without evaluating
the condition fully.

To fix this issue, add a condition to test whether firmware is
down or not using ATH10K_FLAG_CRASH_FLUSH bit after timeout happens.
If ATH10K_FLAG_CRASH_FLUSH bit is also not set then only queue the work
to restart the hardware.

BUG=b:132147953
TEST=None

Change-Id: I0d5b4e842b7dd3e23757803ea822eed518084bea
Signed-off-by: Surabhi Vishnoi <svishnoi@codeaurora.org>
Signed-off-by: Vamsi Singamsetty <vamssi@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1597060
Reviewed-by: Zhihong Yu <zhihongyu@chromium.org>
Reviewed-by: Govind Singh <govinds@qualcomm.corp-partner.google.com>
Commit-Queue: Zhihong Yu <zhihongyu@chromium.org>
Tested-by: Zhihong Yu <zhihongyu@chromium.org>