CHROMIUM: WAR: misc: fastrpc: allocate process release context before hand
In use cases where a fatal signal is pending on the process
either by user SIGINT or a genuine fatal signal.
Allocating CMA memory in such cases would fail with -EINTR as
the process is about to die. However the remote dsp has to
release the init process, to allow this preallocate remote
context to be used in release call.

BUG=b:132147953
TEST=Validate by running label_image -a 1

Change-Id: Idd4d95a78e50a1a4f7c829153dd74e1d6e7055cd
Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@linaro.org>
Signed-off-by: Khasim Syed Mohammed <khasim.mohammed@linaro.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1670753
Commit-Queue: Zhihong Yu <zhihongyu@chromium.org>
Tested-by: Zhihong Yu <zhihongyu@chromium.org>
Reviewed-by: Zhihong Yu <zhihongyu@chromium.org>