misc: fastrpc: workaround DSP access issue
without this change, the DSP might some times fail to respond to the
RELEASE call hence locking the application which will remain blocked.
(it could also be that rpmsg fails to send the message OR fails to
process the callback from the DSP)

In the past the driver was actually failing to send the messages
because using CMA - instead of MMU- the memory needed for the message
could not be allocated due to the pending signal.

However what we observe now using the SMMU is that rpmsg_send will
return success on the RELEASE call but it is the response that never
arrives causing the device to not be released and userspace to
remain  blocked.

Signed-off-by: Jorge Ramirez-Ortiz <jorge.ramirez-ortiz@linaro.org>